+++ src/css/modal.css
@@ -81,6 +81,10 @@
   letter-spacing: 0.5px;
 }
 
+.search-result {
+  overflow-y: scroll;
+}
+
 .video-list {
   width: 1040px;
   height: 493px;
@@ -88,7 +92,6 @@
   flex-direction: row;
   flex-wrap: wrap;
   gap: 32px 20px;
-  overflow-y: scroll;
 }
 
 .video-item {
