+++ src/js/index.js
@@ -2,14 +2,15 @@ import '../css/index.css';
 import '../assets/images/not_found.png';
 import '../assets/images/empty_storage.jpeg';
 import { $ } from './util/dom.js';
-import { MESSAGE, VIDEO } from './constants/constants.js';
-import { isEndOfScroll, throttle } from './util/general.js';
 import storage from './storage/storage.js';
-import YoutubeSearch from './domain/YoutubeSearch.js';
-import searchResultView from './ui/searchResultView.js';
+import searchResultView from './views/searchResultView.js';
+import { handleSearch, handleVideoListScroll } from './handlers/searchModal.js';
+import { handleDeleteVideo, handleWatchedVideo } from './handlers/manageSavedVideo.js';
+import { handleSaveVideos } from './handlers/saveVideo.js';
+import { MESSAGE, VIDEO } from './constants/constants.js';
+import { throttle } from './util/general.js';
 
 export default function App() {
-  const youtubeSearch = new YoutubeSearch();
   const content = 'unseen';
 
   const initSavedVideo = () => {
@@ -20,85 +21,12 @@ export default function App() {
     }
   };
 
-  const renderHandler = async () => {
-    searchResultView.renderSkeletonUI();
-    const response = await youtubeSearch.fetchYoutubeAPI();
-    searchResultView.renderSearchResult(response);
-    const isLastVideos = response.items.length !== 0 && !response.nextPageToken;
-    if (isLastVideos) {
-      $('.video-list').removeEventListener(
-        'scroll',
-        throttle(handleVideoListScroll, VIDEO.THROTTLE_DELAY)
-      );
-    }
-  };
+  $('.search-result').addEventListener(
+    'scroll',
+    throttle(handleVideoListScroll, VIDEO.THROTTLE_DELAY)
+  );
 
-  const handleSearch = () => {
-    try {
-      const searchInput = $('#search-input-keyword').value.trim();
-      youtubeSearch.searchTarget = searchInput;
-      youtubeSearch.pageToken = '';
-      searchResultView.resetVideoList();
-      addVideoListEvents();
-      renderHandler();
-    } catch (error) {
-      alert(error.message);
-    }
-  };
-
-  const handleVideoListScroll = () => {
-    if (isEndOfScroll($('.video-list'))) {
-      renderHandler();
-    }
-  };
-
-  const selectedVideoData = (videoItem) => {
-    const videoData = {
-      videoId: videoItem.dataset.videoId,
-      thumbnails: videoItem.querySelector('.video-item__thumbnail').src,
-      title: videoItem.querySelector('.video-item__title').textContent,
-      channelTitle: videoItem.querySelector('.video-item__channel-name').textContent,
-      publishTime: videoItem.querySelector('.video-item__published-date').textContent,
-      state: 'unseen',
-    };
-    return videoData;
-  };
-
-  const handleSaveVideos = (e) => {
-    const isSaveButtonClick = e.target.classList.contains('video-item__save-button');
-    if (!isSaveButtonClick) {
-      return;
-    }
-    e.target.hidden = true;
-    const videoData = selectedVideoData(e.target.closest('li'));
-    storage.saveVideo(videoData);
-    $('.empty-video-image').classList.add('hide');
-    searchResultView.rendersavedVideo(videoData);
-  };
-
-  const changedVideoList = (changedData) => {
-    storage.setLocalStorage(changedData);
-    searchResultView.renderSavedVideos(content, changedData);
-    if (changedData.length === 0) {
-      storage.resetLocalStorage();
-    }
-  };
-
-  const handleDeleteVideo = (selectedVideoId) => {
-    const deletedData = storage
-      .getLocalStorage()
-      .filter((video) => video.videoId !== selectedVideoId);
-    changedVideoList(deletedData);
-  };
-
-  const addVideoListEvents = () => {
-    $('.video-list').addEventListener(
-      'scroll',
-      throttle(handleVideoListScroll, VIDEO.THROTTLE_DELAY)
-    );
-
-    $('.video-list').addEventListener('click', handleSaveVideos);
-  };
+  $('.search-result').addEventListener('click', handleSaveVideos);
 
   $('.saved-video-list').addEventListener('click', (e) => {
     if (e.target.classList.contains('video-remove-button')) {
@@ -109,14 +37,7 @@ export default function App() {
     }
     if (e.target.classList.contains('video-watched-button')) {
       const selectedVideoId = e.target.closest('li').dataset.videoId;
-      const savedVideos = storage.getLocalStorage();
-      savedVideos.forEach((video) => {
-        if (video.videoId === selectedVideoId) {
-          // eslint-disable-next-line no-param-reassign
-          video.state = 'watched';
-        }
-      });
-      changedVideoList(savedVideos);
+      handleWatchedVideo(selectedVideoId);
     }
   });
