+++ src/js/views/searchResultView.js
